+++ lndtip/Cargo.toml
@@ -11,5 +11,5 @@ tokio = { version = "0.2", features = ["macros","stream"] }
 toml = "0.5"
 serde = { version = "1.0", features = ["derive"] }
 futures = { version = "0.3", default-features = false, features = ["alloc"] }
-r2d2_lndclient = { path="../r2d2_lndclient"}
+mobc_lndclient = { path="../mobc_lndclient"}
 base64 = "0.12.0"
