+++ lndtip/src/lnd_service.rs
@@ -1,6 +1,6 @@
 use base64;
-use r2d2_lndclient::lnd_pool::LightningConnectionManager;
-use r2d2_lndclient::r2d2::Pool;
+use mobc_lndclient::LightningConnectionManager;
+use mobc_lndclient::mobc::Pool;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone)]
@@ -13,10 +13,10 @@ impl LightningService {
         Self { lnc: lnc }
     }
 
-    pub fn add_invoice(&self, satoshi: i64, description: &str, expiry: i64) -> InvoiceResponse {
-        let conn = self.lnc.get().unwrap();
-        if let Ok(invoice_response) = conn.add_invoice(satoshi, description, expiry) {
-            if let Ok(invoice) = conn.lookup_invoice(invoice_response.r_hash.as_slice()) {
+    pub async fn add_invoice(&self, satoshi: i64, description: &str, expiry: i64) -> InvoiceResponse {
+        let mut conn = self.lnc.get().await.unwrap();
+        if let Ok(invoice_response) = conn.add_invoice(satoshi, description, expiry).await {
+            if let Ok(invoice) = conn.lookup_invoice(invoice_response.r_hash.as_slice()).await {
                 let r_hash = base64::encode(invoice.r_hash);
                 return InvoiceResponse {
                     r_hash: r_hash,
@@ -32,10 +32,10 @@ impl LightningService {
         }
     }
 
-    pub fn lookup(&self, r_hash: &str) -> InvoiceStatus {
-        let conn = self.lnc.get().unwrap();
+    pub async fn lookup(&self, r_hash: &str) -> InvoiceStatus {
+        let mut conn = self.lnc.get().await.unwrap();
         if let Ok(r_hash) = base64::decode(r_hash) {
-            if let Ok(invoice) = conn.lookup_invoice(r_hash.as_slice()) {
+            if let Ok(invoice) = conn.lookup_invoice(r_hash.as_slice()).await {
                 return InvoiceStatus {
                     status: format!("{:?}", invoice.state),
                     settled: invoice.settled,
