+++ mobc_lndclient/README.md
@@ -1,0 +1,3 @@
+Links
+
+* https://blog.logrocket.com/create-an-async-crud-web-service-in-rust-with-warp/
