--- r2d2_lndclient/README.md
@@ -1,16 +1,0 @@
-# rust rd2d2 lndclient
-
-
-## More info
-
-* https://github.com/sfackler/r2d2
-
-## License
-
-Rust r2d2 lightningnetwork client
-
-Copyright(c) 2020 David Rasch
-
-This program is free software; you can redistribute it and/or modify it under the terms of the GNU General Public License (GPL) as published by the Free Software Foundation; either version 3 of the License, or (at your option) any later version.
-
-This program is distributed in the hope that it will be useful, but without any warranty; without even the implied warranty of merchantability or fitness for a particular purpose. See the GNU General Public License for more details.
