--- r2d2_lndclient/src/lib.rs
@@ -1,3 +1,0 @@
-pub extern crate r2d2;
-pub extern crate rust_lndclient;
-pub mod lnd_pool;
